+++ packages/core/src/schema/reduce.ts
@@ -89,6 +89,7 @@ type ReduceValueForReducer<
 > = Using extends
   | 'sum'
   | 'avg'
+  | 'mean'
   | 'count'
   | 'min'
   | 'max'